+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+* Improved check for `None` arguments in `BBoxes`.
+
 ## [0.1.3] - 2020-04-04
 
 ### Changed