+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.1.5] - 2020-04-25
+
+### Fixed
+
+* Improve type checking for integers to take into account numpy integers.
+
 ## [0.1.4] - 2020-04-24
 
 ### Added